+++ src/Components/Charts/DeviceData/index.js
@@ -1,7 +1,14 @@
 export const deviceData = [
-    { dateTime: '2023-10-15T10:00:00', totalUsers: 50, devicesUsed: 20, activeUsers: 15 },
-    { dateTime: '2023-10-15T12:00:00', totalUsers: 80, devicesUsed: 30, activeUsers: 25 },
-    { dateTime: '2023-10-15T14:00:00', totalUsers: 90, devicesUsed: 40, activeUsers: 35 },
-    
+    { dateTime: '2023-10-15T10:00:00', totalUsers: 10, devicesUsed: 65, activeUsers: 15 },
+    { dateTime: '2023-10-15T12:00:00', totalUsers: 200, devicesUsed: 705, activeUsers: 125 },
+    { dateTime: '2023-10-15T14:00:00', totalUsers: 300, devicesUsed: 65, activeUsers: 5 },
+    { dateTime: '2023-10-15T14:00:00', totalUsers: 40, devicesUsed: 605, activeUsers: 350 },
+    { dateTime: '2023-10-15T14:00:00', totalUsers: 500, devicesUsed: 650, activeUsers: 235 },
+    { dateTime: '2023-10-15T14:00:00', totalUsers: 100, devicesUsed: 40, activeUsers: 735 },
+    { dateTime: '2023-10-15T14:00:00', totalUsers: 10, devicesUsed: 400, activeUsers: 305 },
+    { dateTime: '2023-10-15T14:00:00', totalUsers: 50, devicesUsed: 208, activeUsers: 359 },
+    { dateTime: '2023-10-15T14:00:00', totalUsers: 70, devicesUsed: 400, activeUsers: 535 },
+   
+   
   ];